Add point shadow example.
Now we can render point shadow and sunlight shadow.
But only one point light and sunlight can be used to calculate shadow now. (just for shadow calculating, not shading and creating. It's able to create multiple point lights, but only the first one would be used to calculate shadow in shader.)